Fim de expediente

+++ app/controller/task_controller.py
@@ -30,20 +30,59 @@ def create_task():
         return redirect(url_for('index_bp.tasks_page'))
     
 
-@task_bp.route('/task/<int:task_id>')
+
+@task_bp.route('/delete/<int:task_id>')
 def delete_task(task_id):
     if session ['login'] == None or 'login' not in session:
         return redirect(url_for('index_bp.login_page'))
     
-    deleted = TaskService.delete_task(task_id)
+    find_task = TaskService.find_by_id(task_id)
     
-    if session.get('login'):
+    if find_task:
+        TaskService.delete_task(task_id)
         flash('Tarefa deletada com sucesso!')
         return redirect(url_for('index_bp.tasks_page'))
+    
     else:
         flash('Não foi possível excluir a tarefa!')
         return redirect(url_for('index_bp.tasks_page'))
-        
     
-        
+@task_bp.route('/update', methods=['POST'])
+def update_task():
+
+    if session ['login'] == None or 'login' not in session:
+        return redirect(url_for('index_bp.login_page'))
+
+    user_id = session.get('user_id')
+
+    form = CreateTaskForm(request.form)
+
+    if form.validate_on_submit():
+
+        task_id = request.form.get('inputId')
+
+        task = TaskService.find_by_id(request.form['inputId'])
+        TaskService.update_task(task_id, form, user_id)
+
+        if task:
+            flash('Tarefa editada com sucesso!')
+            return redirect(url_for('index_bp.tasks_page'))
+        else:
+            flash('Não foi possível editar a tarefa!')
+            redirect(url_for('index_bp.update_page'))
             
+@task_bp.route('/done/<int:task_id>')
+def done_task(task_id):
+
+    if session['login'] == None or 'login' not in session:
+        flash('É necessário se autenticar!')
+        return redirect(url_for('index_bp.login_page'))
+
+    sucess = TaskService.update_status(task_id)
+
+    if sucess:
+        flash('Tarefa finalizada!')
+        return redirect(url_for('index_bp.tasks_page'))   
+    else: 
+        flash('Não foi possível atualizar o status da tarefa!')
+        return redirect(url_for('index_bp.tasks_page'))
